requirements.txt
"""Project dependencies and requirements"""

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23

# OCR & Image Processing
pytesseract==0.3.10
pillow==10.1.0

# LLM & AI
ollama==0.1.2
langchain==0.1.0
langgraph==0.0.44

# Logging
loguru==0.7.2

# HTTP Client
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development (Optional)
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
